fix: remove cli & default server conf as struct
