+++ cmd/meta/meta.go
@@ -4,15 +4,54 @@
 package main
 
 import (
+	"encoding/json"
+	"os"
+
 	"github.com/alecthomas/kong"
-	"github.com/tatris-io/tatris/cmd"
+	"github.com/gin-gonic/gin"
+	"github.com/tatris-io/tatris/internal/common/log"
+	"github.com/tatris-io/tatris/internal/common/log/logger"
+	"github.com/tatris-io/tatris/internal/common/log/util"
 	"github.com/tatris-io/tatris/internal/service"
+	"go.uber.org/zap"
 )
 
-var cli cmd.Cli
+var cli struct {
+	Debug bool `help:"Enable debug mode."`
+	Conf  struct {
+		Logging string `type:"existingfile" help:"Logging config file path."`
+		Meta    string `type:"existingfile" help:"Meta config file path."`
+	} `embed:"" prefix:"conf."`
+}
+
+// if init failure occurs, the program will use the default logger
+func initLoggers(confPath string) {
+	content, err := os.ReadFile(confPath)
+	if err != nil {
+		logger.Panic("fail to init loggers, use the default console logger instead", zap.Error(err))
+		return
+	}
+	var logConf util.Config
+	if err := json.Unmarshal(content, &logConf); err != nil {
+		logger.Panic("fail to init loggers, use the default console logger instead", zap.Error(err))
+		return
+	}
+	// validate all confs
+	logConf.Verify()
+	log.InitLoggers(&logConf)
+}
 
 func main() {
 	kong.Parse(&cli)
-	cmd.Initialize(&cli)
+
+	if len(cli.Conf.Logging) != 0 {
+		initLoggers(cli.Conf.Logging)
+	}
+
+	if cli.Debug {
+		gin.SetMode(gin.DebugMode)
+	} else {
+		gin.SetMode(gin.ReleaseMode)
+	}
 	service.StartHTTPServer("meta")
 }
