+++ conf/log-conf.json
@@ -6,7 +6,7 @@
   "global-logger": {
     "file-confs": [
       {
-        "file-name": "global-error.log",
+        "file-name": "error.log",
         "max-size": 300,
         "max-days": 5,
         "max-backups": 5,
@@ -14,7 +14,7 @@
         "level-min": "error"
       },
       {
-        "file-name": "global-all.log",
+        "file-name": "all.log",
         "max-size": 300,
         "max-days": 5,
         "max-backups": 5,
