+++ internal/meta/metadata/index_manager.go
@@ -7,8 +7,9 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
-	"github.com/patrickmn/go-cache"
 	"strings"
+
+	"github.com/patrickmn/go-cache"
 
 	"go.uber.org/zap"
